# openapi/transaction-api.yaml
openapi: 3.1.0
info:
  title: transaction-service public API
  version: 1.0.0
  summary: Send Money (wallet to wallet) for the MFS POC v2.
  description: |
    Authoritative, contract-first definition of the public API of `transaction-service`
    (Send Money POC v2, spec section 7.1). The implementation and its tests MUST match this file;
    it is served unchanged at `GET /openapi.yaml`.

    **Money.** Every amount is an integer number of **poisha** (BDT minor units, 1 BDT = 100 poisha).
    Currency is always `BDT`. Timestamps are UTC, RFC 3339 (`date-time`).

    **Authentication.** Every `/api/**` operation requires the static API key in the `X-API-Key`
    header (NFR-09). A missing or wrong key returns `401 UNAUTHORIZED`.

    **Idempotency.** `POST /api/v1/send-money` and `POST /api/v1/wallets/{msisdn}/fund` require an
    `Idempotency-Key` header (1..64 characters). The same key from the same sender with the same
    body returns the stored result; the same key with a different body returns
    `409 IDEMPOTENCY_CONFLICT`. Wallet registration is idempotent on the MSISDN.

    **Outcomes of Send Money.**

    | HTTP | Meaning | Body |
    | --- | --- | --- |
    | 200 | `COMPLETED`: every leg is durably posted in the ledger | `SendMoneyResponse` |
    | 202 | `PROCESSING`: ledger outcome not yet known; poll `GET /api/v1/send-money/{txnId}` | `SendMoneyResponse` |
...
    | 409 | `IDEMPOTENCY_CONFLICT` or `QUOTE_CHANGED` | `Problem` |
...

...

...
  contact:
    name: MFS POC team
...
servers:
...
security:
  - apiKey: []
paths:
...
  /api/v1/send-money:
    post:
      tags: [send-money]
      operationId: sendMoney
      summary: Send Money (synchronous, idempotent)
      description: |
        FR-02. Validates, prices, reserves limits, posts all legs atomically to the ledger and
        returns the final status in the same response.

        - **200** `COMPLETED`: the ledger acknowledged the posting (NFR-04).
        - **202** `PROCESSING`: the ledger outcome is unknown (timeouts after retries, ledger 409/500,
          ledger bulkhead full) or the same `Idempotency-Key` is still in flight. The repair worker
          resolves it; poll `GET /api/v1/send-money/{txnId}`.
        - **422** `FAILED`: a business rule failed (no record is written, no `txnId`) or the ledger
          rejected the posting definitively (record FAILED, limits released, `txnId` present).

        If `quoteToken` is sent, the fee MUST equal the quoted fee and the token must match the
        sender, receiver and amount and be unexpired, otherwise `409 QUOTE_CHANGED`. A malformed
        token or a bad signature is `400 INVALID_QUOTE_TOKEN`.

        A replay (same sender, same `Idempotency-Key`, same body) returns the stored outcome with
        the same status code as the original (200, 202 while in flight, or 422).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMoneyRequest'
            examples:
              rent:
                summary: Worked example (spec 3) with a reference and a quote token
                value:
                  senderMsisdn: '01711000001'
                  receiverMsisdn: '01811000002'
                  amount: 100000
                  currency: BDT
                  reference: Rent
                  quoteToken: eyJzIjoiMDE3MTEwMDAwMDEiLCJyIjoiMDE4MTEwMDAwMDIiLCJhIjoxMDAwMDAsImYiOjUwMCwiZSI6MTc5MTM1MTE0MzIxMX0.Zq3k1w1y3yX1pZ5cO0Jm8xkD3t0z2vI6eV8m3c1aPz9
              minimal:
                summary: Required fields only
                value:
                  senderMsisdn: '01711000001'
                  receiverMsisdn: '01811000002'
                  amount: 100000
                  currency: BDT
      responses:
        '200':
          description: COMPLETED. Every leg is durably posted in the ledger.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMoneyResponse'
              examples:
                completed:
                  value:
                    txnId: 0192f5a4-7c2e-7b1a-9d3e-5f4a2b1c0d00
                    status: COMPLETED
                    amount: 100000
                    fee: 500
                    vat: 65
                    commission: 87
                    totalDebit: 100500
                    completedAt: '2026-10-07T09:14:03.211Z'
        '202':
          description: |
            PROCESSING. The record exists and the outcome is being resolved; poll the status
            endpoint. `completedAt` is absent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMoneyResponse'
              examples:
                processing:
                  value:
                    txnId: 0192f5a4-7c2e-7b1a-9d3e-5f4a2b1c0d00
                    status: PROCESSING
                    amount: 100000
                    fee: 500
                    vat: 65
                    commission: 87
                    totalDebit: 100500
                    message: "The ledger has not confirmed this transfer yet (ledger-service is slow or unavailable). It is completed or failed automatically; poll GET /api/v1/send-money/{txnId}"
        '400':
          description: |
            `VALIDATION_FAILED` (invalid body, missing or too long `Idempotency-Key`) or
            `INVALID_QUOTE_TOKEN` (malformed token or bad signature). Nothing was written.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                validationFailed:
                  $ref: '#/components/examples/ValidationFailedProblem'
                missingIdempotencyKey:
                  value:
                    type: urn:problem-type:mfs:VALIDATION_FAILED
                    title: Bad Request
                    status: 400
                    detail: Required header 'Idempotency-Key' is missing
                    instance: /api/v1/send-money
                    code: VALIDATION_FAILED
                    errors:
                      - field: Idempotency-Key
                        message: must not be blank
                invalidQuoteToken:
                  value:
                    type: urn:problem-type:mfs:INVALID_QUOTE_TOKEN
                    title: Bad Request
                    status: 400
                    detail: The quote token is malformed or its signature is invalid
                    instance: /api/v1/send-money
                    code: INVALID_QUOTE_TOKEN
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: |
            `IDEMPOTENCY_CONFLICT`: the `Idempotency-Key` was already used by this sender with a
            different body (`txnId` of the original is included). `QUOTE_CHANGED`: the quote expired,
            does not match the request, or the fee changed. Do not retry the same request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                idempotencyConflict:
                  value:
                    type: urn:problem-type:mfs:IDEMPOTENCY_CONFLICT
                    title: Conflict
                    status: 409
                    detail: Idempotency-Key was already used with a different request body
                    instance: /api/v1/send-money
                    code: IDEMPOTENCY_CONFLICT
                    txnId: 0192f5a4-7c2e-7b1a-9d3e-5f4a2b1c0d00
                quoteChanged:
                  value:
                    type: urn:problem-type:mfs:QUOTE_CHANGED
                    title: Conflict
                    status: 409
                    detail: The quote has expired or no longer matches; request a new quote
                    instance: /api/v1/send-money
                    code: QUOTE_CHANGED
        '422':
          description: |
            FAILED. `code` is the reason. Rule failures (`SELF_TRANSFER`, `WALLET_NOT_FOUND`,
            `WALLET_INACTIVE`, `RECEIVER_NOT_ALLOWED`, `AMOUNT_OUT_OF_RANGE`, `LIMIT_EXCEEDED`) write
            no record and carry no `txnId`. Ledger rejections (`INSUFFICIENT_FUNDS`,
            `WALLET_NOT_FOUND`, `LEDGER_REJECTED`) mark the record FAILED, release the limits and carry
            its `txnId`. Not retryable as is.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                insufficientFunds:
                  value:
                    type: urn:problem-type:mfs:INSUFFICIENT_FUNDS
                    title: Unprocessable Content
                    status: 422
                    detail: The sender's available balance is too low
                    instance: /api/v1/send-money
                    code: INSUFFICIENT_FUNDS
                    txnId: 0192f5a4-7c2e-7b1a-9d3e-5f4a2b1c0d00
                limitExceeded:
                  value:
                    type: urn:problem-type:mfs:LIMIT_EXCEEDED
                    title: Unprocessable Content
                    status: 422
                    detail: The sender's daily or monthly limit would be exceeded
                    instance: /api/v1/send-money
                    code: LIMIT_EXCEEDED
                selfTransfer:
                  $ref: '#/components/examples/SelfTransferProblem'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static API key (environment variable `API_KEY`, property `poc.security.api-key`).
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-chosen key, unique per sender and logical operation (for example a UUID). Re-sending
        the same key with the same body returns the stored result.
      schema:
        type: string
        minLength: 1
        maxLength: 64
      example: 7f9c2ba4-e88f-4d2a-9d3b-1c5e8a7f6b21
...
  headers:
    RetryAfter:
      description: Seconds to wait before retrying.
      required: true
      schema:
        type: integer
        minimum: 1
      example: 1
  examples:
    ValidationFailedProblem:
      value:
        type: urn:problem-type:mfs:VALIDATION_FAILED
        title: Bad Request
        status: 400
        detail: Request validation failed
        instance: /api/v1/send-money
        code: VALIDATION_FAILED
        errors:
          - field: amount
            message: must be greater than 0
          - field: receiverMsisdn
            message: must match "^01[3-9][0-9]{8}$"
    SelfTransferProblem:
      value:
        type: urn:problem-type:mfs:SELF_TRANSFER
        title: Unprocessable Content
        status: 422
        detail: Sender and receiver must be different wallets
        instance: /api/v1/send-money
        code: SELF_TRANSFER
...
  responses:
...
    Unauthorized:
      description: '`UNAUTHORIZED`: the `X-API-Key` header is missing or wrong.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unauthorized:
              value:
                type: urn:problem-type:mfs:UNAUTHORIZED
                title: Unauthorized
                status: 401
                detail: Missing or invalid API key
                instance: /api/v1/send-money
                code: UNAUTHORIZED
...
    Unavailable:
      description: |
        `OVERLOADED` (a bulkhead is saturated, the request was shed before any write),
        `LEDGER_UNAVAILABLE` (ledger-service, a separate service, is unreachable: Send Money is refused
        before any write, so no money moved) or `SERVICE_UNAVAILABLE` (PostgreSQL is unavailable
        before any money moved).
        Retry the identical request (same `Idempotency-Key`) after `Retry-After` seconds.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            overloaded:
              value:
                type: urn:problem-type:mfs:OVERLOADED
                title: Service Unavailable
                status: 503
                detail: Too many concurrent requests; retry later
                instance: /api/v1/send-money
                code: OVERLOADED
            serviceUnavailable:
              value:
                type: urn:problem-type:mfs:SERVICE_UNAVAILABLE
                title: Service Unavailable
                status: 503
                detail: A dependency is unavailable; no money was moved
                instance: /api/v1/send-money
                code: SERVICE_UNAVAILABLE
            ledgerUnavailable:
              value:
                type: urn:problem-type:mfs:ledger_unavailable
                title: Service Unavailable
                status: 503
                detail: ledger-service is unavailable. No money was moved; retry after the Retry-After delay
                instance: /api/v1/send-money
                code: LEDGER_UNAVAILABLE
    InternalError:
      description: '`INTERNAL_ERROR`: unexpected error. For Send Money, poll the status before retrying.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            internalError:
              value:
                type: urn:problem-type:mfs:INTERNAL_ERROR
                title: Internal Server Error
                status: 500
                detail: Unexpected error
                instance: /api/v1/send-money
                code: INTERNAL_ERROR
  schemas:
    Msisdn:
      type: string
      pattern: '^01[3-9][0-9]{8}$'
      description: Bangladeshi mobile number (11 digits, operator prefix 013–019).
      examples: ['01711000001']
    Currency:
      type: string
      pattern: '^BDT$'
      description: ISO 4217 currency code. Only `BDT` is supported.
      examples: [BDT]
    Poisha:
      type: integer
      format: int64
      description: Amount in poisha (BDT minor units).
    PositivePoisha:
      type: integer
      format: int64
      minimum: 1
      description: Amount in poisha (BDT minor units), greater than 0.
    TxnStatus:
      type: string
      enum: [PROCESSING, COMPLETED, FAILED]
      description: API status. A record still `INITIATED` in the database is shown as `PROCESSING`.
...
    ErrorCode:
      type: string
      description: |
        The `code` member of every problem: a business `FailureCode` or an API error code.

        | Code | HTTP |
        | --- | --- |
        | `INSUFFICIENT_FUNDS`, `LIMIT_EXCEEDED`, `AMOUNT_OUT_OF_RANGE`, `SELF_TRANSFER`, `WALLET_NOT_FOUND`, `WALLET_INACTIVE`, `RECEIVER_NOT_ALLOWED`, `LEDGER_REJECTED` | 422 |
        | `QUOTE_CHANGED`, `IDEMPOTENCY_CONFLICT`, `MSISDN_EXISTS` | 409 |
        | `VALIDATION_FAILED`, `INVALID_QUOTE_TOKEN` | 400 |
        | `UNAUTHORIZED` | 401 |
        | `NOT_FOUND` | 404 |
        | `OVERLOADED`, `LEDGER_UNAVAILABLE`, `SERVICE_UNAVAILABLE` | 503 |
        | `INTERNAL_ERROR` | 500 |
      enum:
        - INSUFFICIENT_FUNDS
        - LIMIT_EXCEEDED
        - AMOUNT_OUT_OF_RANGE
        - SELF_TRANSFER
        - WALLET_NOT_FOUND
        - WALLET_INACTIVE
        - RECEIVER_NOT_ALLOWED
        - QUOTE_CHANGED
        - IDEMPOTENCY_CONFLICT
        - LEDGER_REJECTED
        - VALIDATION_FAILED
        - INVALID_QUOTE_TOKEN
        - UNAUTHORIZED
        - NOT_FOUND
        - MSISDN_EXISTS
        - OVERLOADED
        - SERVICE_UNAVAILABLE
        - LEDGER_UNAVAILABLE
        - INTERNAL_ERROR
...
    SendMoneyRequest:
      type: object
      required: [senderMsisdn, receiverMsisdn, amount, currency]
      properties:
        senderMsisdn:
          $ref: '#/components/schemas/Msisdn'
        receiverMsisdn:
          $ref: '#/components/schemas/Msisdn'
        amount:
          $ref: '#/components/schemas/PositivePoisha'
        currency:
          $ref: '#/components/schemas/Currency'
        reference:
          type: [string, 'null']
          maxLength: 50
          description: Free-text reference shown to both parties.
          examples: [Rent]
        quoteToken:
          type: [string, 'null']
          maxLength: 512
          description: Optional token from the quote endpoint; locks the fee.
    SendMoneyResponse:
      type: object
      description: Body of 200 COMPLETED and 202 PROCESSING.
      required: [txnId, status, amount, fee, vat, commission, totalDebit]
      properties:
        txnId:
          type: string
          format: uuid
          description: Time-ordered transaction id (low byte 00).
        status:
          $ref: '#/components/schemas/TxnStatus'
        amount:
          $ref: '#/components/schemas/Poisha'
        fee:
          $ref: '#/components/schemas/Poisha'
        vat:
          $ref: '#/components/schemas/Poisha'
        commission:
          $ref: '#/components/schemas/Poisha'
        totalDebit:
          allOf:
            - $ref: '#/components/schemas/Poisha'
          description: amount + fee.
        completedAt:
          type: [string, 'null']
          format: date-time
          description: Set when COMPLETED; omitted while PROCESSING.
        message:
          type: [string, 'null']
          description: Set while PROCESSING to explain that the ledger has not confirmed the transfer yet; omitted otherwise.
...
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          examples: [amount]
        message:
          type: string
          examples: [must be greater than 0]
    Problem:
      type: object
      description: RFC 9457 Problem Details with the extension members `code`, `txnId` and `errors`.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
          description: '`urn:problem-type:mfs:<code>` (or `about:blank` for framework-generated problems).'
          examples: ['urn:problem-type:mfs:INSUFFICIENT_FUNDS']
        title:
          type: string
        status:
          type: integer
          format: int32
          description: The HTTP status code.
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          $ref: '#/components/schemas/ErrorCode'
        txnId:
          type: string
          format: uuid
          description: Present when a transaction record exists (ledger rejection, idempotency conflict).
        errors:
          type: array
          description: Field-level validation errors (400 only).
          items:
            $ref: '#/components/schemas/FieldError'