# src/main/java/com/bracits/transactionservice/api/dto/response/FieldError.java
package com.bracits.transactionservice.api.dto.response;

/**
 * One item of the Problem Details {@code errors} member: the offending field (or parameter) and
 * why.
 */
public record FieldError(String field, String message) {

}